+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.4.3
+
+- Added the missing test cases for EDULandRewards._claimRewards() where some batchNumber is 0.
+- Updated the MockReferee contract to achieve the above test case.
+- Further removed duplicated _msgSender() call in EDULand.
+
+## 1.4.2
+
+- Remediations based on internal audit AB-SC-SOL-01 to AB-SC-SOL-09
+- Expose EDULandRewards.isKycWallet() public view function
+- Skip but not revert if the recipient of a claimable batch in EDULandRewards hasn't been registered as a KYC wallet.
+
 ## 1.4.1
 
 - Revert with TokenAlreadyRented error in EDULandRental.estimateRentalFee() view function if expired tokens are not collected